refactor(webui): isolate desktop data boundary
