+++ package.json
@@ -16,7 +16,7 @@
     "lint": "eslint \"frontend/dashboard/src/**/*.{ts,tsx}\" \"frontend/chat/src/**/*.{ts,tsx}\" \"plugins/**/*.ts\" \"types/**/*.d.ts\"",
     "package:mobile-web": "bash scripts/package-mobile-web.sh",
     "test:shared-chat-showcase": "node --test frontend/chat/src/shared-chat-showcase.test.mjs frontend/chat/src/trace-motion-showcase.test.mjs frontend/chat/src/drawer-island-showcase.test.mjs",
-    "test:desktop-navigation": "node --test frontend/chat/src/desktop-navigation.test.mjs frontend/chat/src/desktop-composer.test.mjs frontend/chat/src/model-capsule-data.test.mjs frontend/chat/src/mobile-pairing.test.mjs frontend/chat/src/settings-interactions.test.mjs frontend/chat/src/memory-settings.test.mjs frontend/chat/src/webui-error-boundary.test.mjs",
+    "test:desktop-navigation": "node --experimental-strip-types --test frontend/chat/src/desktop-navigation.test.mjs frontend/chat/src/desktop-composer.test.mjs frontend/chat/src/model-capsule-data.test.mjs frontend/chat/src/mobile-pairing.test.mjs frontend/chat/src/settings-interactions.test.mjs frontend/chat/src/memory-settings.test.mjs frontend/chat/src/web-chat-data.test.mjs frontend/chat/src/webui-error-boundary.test.mjs",
     "test:theme": "node --test frontend/theme/src/theme-catalog.test.mjs",
     "test:dashboard-api": "node --experimental-strip-types --test frontend/dashboard/src/api.test.mjs",
     "test:mobile-web-state": "node --experimental-strip-types --test frontend/chat/src/desktop-conversation.test.mjs frontend/chat/src/static-message-response.test.mjs frontend/chat/src/runtime-dashboard-data.test.mjs frontend/chat/src/message-rendering-policy.test.mjs frontend/chat/src/mobile-message-state.test.mjs frontend/chat/src/mobile-stream-projection.test.mjs frontend/chat/src/web-stream-projection.test.mjs frontend/chat/src/web-chat-status.test.mjs frontend/chat/src/mobile-plugin-layout.test.mjs frontend/chat/src/mobile-plugin-query-queue.test.mjs frontend/chat/src/mobile-plugin-result-cache.test.mjs frontend/chat/src/mobile-surface-history.test.mjs frontend/chat/src/mobile-turn-trace.test.mjs frontend/chat/src/web-turn-trace.test.mjs",
